reworks content intro component, styling, composable for link animation
