+++ composables/useAnimateLink.js
@@ -1,0 +1,46 @@
+import { getJapaneseWord } from '@/lib/weaponsAnimation.js'
+
+export const useAnimateLink = (contentEl) => {
+  const linkText = ref(null)
+  const linkInterval = ref(null)
+  const links = ref([])
+
+  onMounted(() => {
+    links.value = contentEl.value.querySelectorAll('a.link')
+    links.value.forEach((link) => {
+      link.addEventListener('mouseenter', animateLink)
+
+      link.addEventListener('mouseleave', stopAnimateLink)
+    })
+  })
+
+  onBeforeUnmount(() => {
+    links.value.forEach((link) => {
+      link.removeEventListener('mouseenter', animateLink)
+      link.removeEventListener('mouseLeave', stopAnimateLink)
+    })
+  })
+
+  function animateLink(e) {
+    const target = e.currentTarget
+    linkText.value = target.innerText
+
+    setLinkText(target)
+    linkInterval.value = setInterval(() => {
+      setLinkText(target)
+    }, 50)
+  }
+
+  function stopAnimateLink(e) {
+    const target = e.currentTarget
+    clearTimeout(linkInterval.value)
+    target.innerText = linkText.value
+    linkInterval.value = null
+    linkText.value = null
+  }
+
+  function setLinkText(target) {
+    const japWord = getJapaneseWord(target.innerText)
+    target.innerText = japWord
+  }
+}
