+++ pages/experiences/[experience].vue
@@ -3,11 +3,11 @@
     ref="content"
     class="content"
   >
-    <ExperienceIntro
+    <ContentIntro
       :logo-img="experience.logoImg"
       :logo-title="experience.logoTitle"
       :period="experience.period"
-      :stack="experience.stack"
+      :tags="experience.stack"
     />
     <ContentRenderer
       class="content__experience"
@@ -19,7 +19,7 @@
 
 <script setup>
 import { useRoute } from 'vue-router'
-import { getJapaneseWord } from '@/lib/weaponsAnimation.js'
+import { useAnimateLink } from '@/composables/useAnimateLink.js'
 
 const route = useRoute()
 const { data: experience } = await useAsyncData('experience', () =>
@@ -27,48 +27,7 @@ const { data: experience } = await useAsyncData('experience', () =>
 )
 
 const contentEl = useTemplateRef('content')
-const linkText = ref(null)
-const linkInterval = ref(null)
-const links = ref([])
-
-onMounted(() => {
-  links.value = contentEl.value.querySelectorAll('a.link')
-  links.value.forEach((link) => {
-    link.addEventListener('mouseenter', animateLink)
-
-    link.addEventListener('mouseleave', stopAnimateLink)
-  })
-})
-
-onBeforeUnmount(() => {
-  links.value.forEach((link) => {
-    link.removeEventListener('mouseenter', animateLink)
-    link.removeEventListener('mouseLeave', stopAnimateLink)
-  })
-})
-
-function animateLink(e) {
-  const target = e.currentTarget
-  linkText.value = target.innerText
-
-  setLinkText(target)
-  linkInterval.value = setInterval(() => {
-    setLinkText(target)
-  }, 50)
-}
-
-function stopAnimateLink(e) {
-  const target = e.currentTarget
-  clearTimeout(linkInterval.value)
-  target.innerText = linkText.value
-  linkInterval.value = null
-  linkText.value = null
-}
-
-function setLinkText(target) {
-  const japWord = getJapaneseWord(target.innerText)
-  target.innerText = japWord
-}
+useAnimateLink(contentEl)
 </script>
 
 <style scoped lang="scss">
